ai_services/pyproject.toml
[project]
name = "project-root"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "basicsr>=1.4.2",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "groq>=0.36.0",
    "langchain-groq>=1.1.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "paddleocr>=3.3.2",
    "paddlepaddle>=3.2.2",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "realesrgan>=0.3.0",
    "requests>=2.32.5",
    "torch>=2.9.1",
    "werkzeug>=3.1.3",
]
